+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Add `BRIDGE_TO_CCIP` action to `BridgeSettlerBase` for Chainlink CCIP cross-chain token transfers
 * Deploy Settler to Tempo mainnet chain
+  * Add `POSITIVE_SLIPPAGE` action on Tempo
 * Add Thena UniV3 Algebra style fork to Bnb with fork ID 37
 * EkuboV2 VIP actions `EKUBO_VIP` and `METATXN_EKUBO_VIP` are not
   supported anymore and were removed from `ISettlerActions.sol`.